修改sdbimport.log日志不存在时不报错，修改原因：新的CI用例随机调配，第一个跑的用例在执行sdbimprt命令前sdbimport.log是不存在，此时执行mv备份sdbimport.log日志会报错。（预期不需要报错）
git-svn-id: http://192.168.20.11/sequoiadb/trunk@26089 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ testcase_new/story/js/sdbimprt/sdbimprt_fieldsName_5430.js
@@ -40,8 +40,12 @@ function importData( csName, clName, imprtFile )
    println("\n---Begin to import data and check exec result.");
    
    //backup sdbimport.log and generate new sdbimport.log
-   var time = cmd.run( 'date "+%Y-%m-%d-%H:%M:%S"' );
-   cmd.run( 'mv ./sdbimport.log sdbimport.log_'+ time );
+   var rt = cmd.run( 'find ./ -name "sdbimport.log"' );
+   if( rt !== '' )  //file is exist
+   {
+      var time = cmd.run( 'date "+%Y-%m-%d-%H:%M:%S"' );
+      cmd.run( 'mv ./sdbimport.log sdbimport.log_'+ time );
+   }
    
    //---------------------scene1-------------------------
    println( '-----------scene1--Begin to import, headerline exist duplicate field name: a-----------\n' );
